Fixes issue with inboxes and built files
When building the app NOT in dev-mode, the path for checking whether or
not we've bootstrapped the initial files was too broad, so I made it
more specific. I also fixed an issues where newly created cards were
showing up even when the inbox was in collapsed view and the card
indicator wasn't updating.